dotNET packaging: switch from using nupkgs to using pkg-config for building
xbuild looks for .pc files in the pkg-config path of the same names as
the dlls. We use this mechanism to locate DLLs that are required at
build time.

When downloading nupks, we now unzip them and create the appropriate pc
files.

paket is now build using this mechanism. Tests no longer build for some
reason (so I disabled them) but the nix script is now a lot simpler.
patches to project files were factored out.

MonoDevelop nix file was minimally fixed to accomodate the new mechanism
thought it isn't yet taking full advantage of it.