avoid copying files in makeInitrdNG
The makeInitrdNG program previously was creating symlinks and copying
files to a temporary output directory. This was needed because cpio
needed all files to be available under the same directory.

Nix creates the symlinks using `linkFarm`, which results in a directory
representing the eventual initrd root. /nix is is 'bind-mounted' into
this directory using proot, so that cpio can include nix files as well.

With this it is possible for makeInitrdNG to only output paths of files
without making copies. The output of paths is piped into cpio.

This makes the process of building an initrd a tad bit faster.